app/requirements.txt
build==1.2.1
pytest==8.3.2

google-cloud-logging==3.11.1
